refactor country and organisation cache generator
