+++ tasks/data_file_cache/country_recache.py
@@ -1,88 +1,48 @@
 import json
 import os
-import shutil
-import time
-import zipfile
 
 from celery import shared_task
 from django.conf import settings
-from django.db.models import Q
-from openpyxl import load_workbook
 
 from gwml2.models.download_request import WELL_AND_MONITORING_DATA, GGMN
 from gwml2.models.general import Country
 from gwml2.models.well import Well
-from gwml2.tasks.data_file_cache.base_cache import WellCacheFileBase
+from gwml2.tasks.data_file_cache.base_cache import WellCacheZipFileBase
 
-GWML2_FOLDER = settings.GWML2_FOLDER
-DATA_FOLDER = os.path.join(GWML2_FOLDER, 'data')
-WELL_FOLDER = os.path.join(GWML2_FOLDER, 'wells-data')
-DJANGO_ROOT = os.path.dirname(
-    os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-)
-TEMPLATE_FOLDER = os.path.join(DJANGO_ROOT, 'static', 'download_template')
+COUNTRY_DATA_FOLDER = os.path.join(settings.GWML2_FOLDER, 'country-data')
 
 
-class GenerateCountryCacheFile(WellCacheFileBase):
+class GenerateCountryCacheFile(WellCacheZipFileBase):
     @property
     def country(self) -> Country:
         """Return country."""
         return self.country_data
 
     @property
-    def folder(self) -> str:
-        """Return folder.."""
-        return os.path.join(DATA_FOLDER, str(self.country.code))
+    def cache_type(self) -> str:
+        """Return type of cache (country or organisation)."""
+        return 'country'
+
+    @property
+    def cache_name(self) -> str:
+        """Return name of cache object."""
+        return self.country.code
+
+    def get_well_queryset(self):
+        return Well.objects.select_related(
+            'organisation').filter(country=self.country).order_by('-id')
 
     def __init__(self, country):
         self.country_data = country
-        self.current_time = time.time()
-        self.log(f'----- Begin cache country : {country.code}  -------')
+        self.data_folder = COUNTRY_DATA_FOLDER
 
-        # Prepare files
-        well_folder = self.folder_by_type(WELL_AND_MONITORING_DATA)
-        if not os.path.exists(well_folder):
-            os.makedirs(well_folder)
-
-        ggmn_folder = self.folder_by_type(GGMN)
-        if not os.path.exists(ggmn_folder):
-            os.makedirs(ggmn_folder)
-
-        # copy files
-        self.copy_template(self.wells_filename)
-        self.copy_template(self.drill_filename)
-
-        # Get data
-        # Well files
-        well_file = self.file_by_type(
-            self.wells_filename, WELL_AND_MONITORING_DATA)
-        well_book = load_workbook(well_file)
-        well_ggmn_file = self.file_by_type(self.wells_filename, GGMN)
-        well_ggmn_book = load_workbook(well_ggmn_file)
-
-        # Drilling files
-        drilling_file = self.file_by_type(
-            self.drill_filename, WELL_AND_MONITORING_DATA)
-        drilling_book = load_workbook(drilling_file)
-        drilling_ggmn_file = self.file_by_type(self.drill_filename, GGMN)
-        drilling_ggmn_book = load_workbook(drilling_ggmn_file)
-
-        # Save the data
+    def run(self):
+        super(GenerateCountryCacheFile, self).run()
+        # generate organisations json file
         organisations = []
         ggmn_organisations = []
-        wells = Well.objects.filter(country=self.country).order_by('-id')
+        wells = self.get_well_queryset()
         for well in wells:
-            self.merge_data_per_well(
-                well, self.wells_filename, well_book,
-                well_ggmn_book if well.number_of_measurements > 0 and well.organisation else None,
-                ['General Information', 'Hydrogeology', 'Management']
-            )
-            self.merge_data_per_well(
-                well, self.drill_filename, drilling_book,
-                drilling_ggmn_book if well.number_of_measurements > 0 and well.organisation else None,
-                ['Drilling and Construction', 'Water Strike',
-                 'Stratigraphic Log', 'Structures']
-            )
             if (
                 well.organisation and
                 well.organisation.name not in organisations
@@ -93,111 +53,13 @@ class GenerateCountryCacheFile(WellCacheFileBase):
                 well.organisation.name not in ggmn_organisations
             ):
                 ggmn_organisations.append(well.organisation.name)
-
-        # Save book
-        well_book.save(well_file)
-        well_ggmn_book.save(well_ggmn_file)
-        drilling_book.save(drilling_file)
-        drilling_ggmn_book.save(drilling_ggmn_file)
-        # save organisations for a country
         self.generate_organisations_json_file(WELL_AND_MONITORING_DATA,
                                               organisations)
         self.generate_organisations_json_file(GGMN, ggmn_organisations)
 
-        # -------------------------------------------------------------------------
-        # zipping files
-        # -------------------------------------------------------------------------
-        self.log(f'----- Finish constructing country : {country.code}  ------')
-        for data_type in [WELL_AND_MONITORING_DATA, GGMN]:
-            zip_filename = f'{str(self.country.code)} - {data_type}.zip'
-            zip_file = os.path.join(DATA_FOLDER, zip_filename)
-            if os.path.exists(zip_file):
-                os.remove(zip_file)
-            zip_file = zipfile.ZipFile(zip_file, 'w')
-
-            well_file = self.file_by_type(self.wells_filename, data_type)
-            zip_file.write(
-                well_file, self.wells_filename,
-                compress_type=zipfile.ZIP_DEFLATED)
-
-            drill_file = self.file_by_type(self.drill_filename, data_type)
-            zip_file.write(
-                drill_file,
-                self.drill_filename,
-                compress_type=zipfile.ZIP_DEFLATED)
-
-            original_ids_found = {}
-            for well in wells:
-                if well.number_of_measurements == 0:
-                    continue
-                well_folder = os.path.join(WELL_FOLDER, f'{well.id}')
-                measurement_file = os.path.join(
-                    well_folder, self.monitor_filename
-                )
-                if os.path.exists(measurement_file):
-                    original_id = well.original_id
-                    try:
-                        _filename = (
-                            f'monitoring/{original_id} '
-                            f'({original_ids_found[original_id] + 1}).xlsx'
-                        )
-                        continue
-                    except KeyError:
-                        _filename = f'monitoring/{original_id}.xlsx'
-                        original_ids_found[original_id] = 0
-
-                    zip_file.write(
-                        measurement_file,
-                        _filename,
-                        compress_type=zipfile.ZIP_DEFLATED
-                    )
-            zip_file.close()
-        shutil.rmtree(self.folder)
-        self.log(f'----- Finish zipping : {country.code}  -------')
-
-    def merge_data_per_well(
-            self, well, filename, well_book, ggmn_book, sheets
-    ):
-        """Merge data per well.."""
-        well_folder = os.path.join(WELL_FOLDER, f'{well.id}')
-        for sheetname in sheets:
-            self.merge_data_between_sheets(
-                os.path.join(well_folder, filename),
-                well_book, ggmn_book, sheetname
-            )
-        well_book.active = 0
-        if ggmn_book:
-            ggmn_book.active = 0
-
-    def merge_data_between_sheets(
-            self, source_file, target_book, target_book_2, sheetname
-    ):
-        """Merge data between sheets"""
-        if not os.path.exists(source_file) or not target_book:
-            return
-        source_file = os.path.join(source_file, sheetname + '.json')
-        data = []
-        if os.path.exists(source_file):
-            _file = open(source_file, "r")
-            data = json.loads(_file.read())
-
-        # Target book 1
-        target_sheet = target_book[sheetname]
-
-        # Target book 2
-        target_sheet_2 = None
-        if target_book_2:
-            target_sheet_2 = target_book_2[sheetname]
-
-        # Append data from source
-        for row in data:
-            target_sheet.append(row)
-            if target_book_2:
-                target_sheet_2.append(row)
-
     def generate_organisations_json_file(self, data_type, organisations):
-        _file = os.path.join(DATA_FOLDER,
-                             f'{str(self.country.code)} - {data_type}.json')
+        _file = os.path.join(self.data_folder,
+                             f'{str(self.cache_name)} - {data_type}.json')
         with open(_file, "w") as outfile:
             outfile.write(json.dumps(organisations, indent=4))
 
@@ -205,7 +67,8 @@ class GenerateCountryCacheFile(WellCacheFileBase):
 @shared_task(bind=True, queue='update')
 def generate_data_country_cache(self, country_code: str):
     try:
-        country = Country.objects.get(Q(code__iexact=country_code))
-        GenerateCountryCacheFile(country)
+        country = Country.objects.get(code__iexact=country_code)
+        generator = GenerateCountryCacheFile(country)
+        generator.run()
     except Country.DoesNotExist:
         print('Country not found')
